Modififcata la classe DatabaseManager
Ora i metodi di interazione del databse interagiscono con il db.

+++ src/web-api/DatabaseModels/Aim.cs
@@ -7,9 +7,9 @@ namespace WomPlatform.Web.Api.DatabaseModels
 {
     public class Aim
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Description { get; set; }
-        public int ContactId { get; set; }
+        public long ContactId { get; set; }
         public string Type { get; set; }
         public DateTime CreationDate { get; set; }
     }
